pyproject.toml
[project]
name = "hyundai-mqtt"
version = "0.1.0"
description = "MQTT-driven integration with Hyundai Bluelink API for battery/EV data automation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "hyundai-kia-connect-api>=3.0.0",
    "paho-mqtt>=1.6.1",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
